% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{Pokemon data}
\format{
\subsection{A dataframe with 800 rows and 13 columns:}{

\describe{
\item{X.}{index number}
\item{Name}{The pokemon name}
\item{Type.1, Type.2}{The pokemon types}
\item{Type.1, Type.2}{The pokemon types}
\item{Total}{All stats numbers summed}
\item{HP}{HP stats}
\item{Attack}{Attack stats}
\item{Defense}{Defense stats}
\item{Sp..Atk}{Special attack stats}
\item{Sp..Def}{Special Defense stats}
\item{Speed}{Speed stats}
\item{Generation}{pokemon generation}
\item{Legendary}{logical operator indicating if the pokemon is legendary}
}
}
}
\source{
\url{https://gist.github.com/armgilles/194bcff35001e7eb53a2a8b441e8b2c6}
}
\usage{
pokemon
}
\description{
A dataset containing pokemon data from around 2016
}
\keyword{datasets}
