% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_csv_rds.R
\name{tw_csv_rds}
\alias{tw_csv_rds}
\title{creating csv and rds files}
\usage{
tw_csv_rds(x, path)
}
\arguments{
\item{x}{Dataframe or tibble}

\item{path}{The path where the files must be stored. The path must be specified starting at the rproject directory. Do not add .csv or .rds at the end of the file name.}
}
\value{
A .csv and .tsv from a dataframe or tibble at the chosen location
}
\description{
Exports a variable containing a dataframe or tibble into a .csv and .rds file at a chosen location. The packages relies on the readr and here package. A project must be loaded in order to make this command work.
}
\examples{
# A Rproject must be loaded. The Rproject location is the startpoint. Do not add .csv or .rds at the end of the file name.
library(twPackage)
tw_csv_rds(pokemon, path = "pokemon")
}
