% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_factor.R
\name{tw_factor}
\alias{tw_factor}
\title{Make factor}
\usage{
tw_factor(x)
}
\arguments{
\item{x}{The column to be changed into a factor}
}
\value{
Column with type factor
}
\description{
This command changes a column type to factor with unique levels using the unique() command
}
\examples{
library(twPackage)
pokemon$Type.1 <- tw_factor(pokemon$Type.1)
levels(pokemon$Type.1)
}
