% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_filter_select.R
\name{tw_filter_select}
\alias{tw_filter_select}
\title{Filtering and selecting}
\usage{
tw_filter_select(x, filter_col, filter_value, select)
}
\arguments{
\item{x}{the table to be filtered}

\item{filter_col}{column where the filter_value can be found}

\item{filter_value}{value to be filtered on}

\item{select}{the columns to be selected}
}
\value{
A filtered and selected data
}
\description{
This function can be used to filter rows and select columns in one function using the dplyr package
}
\examples{
library(twPackage)
tw_filter_select(
pokemon,
filter_col = "Type.1",
filter_value = "Fire",
select = c("Name", "Type.1")
)
}
