% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_table.R
\name{tw_table}
\alias{tw_table}
\title{tw table}
\usage{
tw_table(x, pagesize)
}
\arguments{
\item{x}{A variable containing a dataframe or tibble}

\item{pagesize}{Number of rows per page}
}
\value{
A basic table made with the reactable package
}
\description{
This command creates a basic table with the reactable package.
}
\examples{
library(twPackage)
tw_table(pokemon, pagesize = 5)
}
