% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twPackage-package.R
\docType{package}
\name{twPackage-package}
\alias{twPackage}
\alias{twPackage-package}
\title{twPackage: Package consisting of 4 functions for a DataScience portofolio}
\description{
The package consists of 4 functions for a DataScience portofolio. "tw_factor" to change a column into a factor. "tw_csv_rds" to export a dataframe/tibble as both an .csv file and .rds file. "tw_table" to make a basic table using the reactable package. "tw_filter_select" to filter rows select columns in one function. The twPackage also contains a pokemon dataset for example purposes.
}
\author{
\strong{Maintainer}: Thijmen Weijgertze \email{thijmen.lifesciences@gmail.com}

}
\keyword{internal}
